+++ Products/CMFPlone/static/plone-compiled.js
@@ -18489,5 +18489,5 @@ require([
 
 });
 
-define("/usr/local/p5dev/buildout.coredev/src/Products.CMFPlone/Products/CMFPlone/static/plone.js", function(){});
+define("/Users/nathan/code/coredev5/src/Products.CMFPlone/Products/CMFPlone/static/plone.js", function(){});